Add schema mode check
Move alias related tests to separate module
